moving folders and adding basic navbar
