+++ src/components/navbar.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+export default function NavBar() {
+  return (
+    <nav class="navbar navbar-expand-lg navbar-light bg-light">
+      <a class="navbar-brand" href="#">
+        Vidly
+      </a>
+    </nav>
+  );
+}
